requirements.txt
altgraph==0.17
asgiref==3.3.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==21.2.0
auditlog3==1.0.1
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
colorama==0.4.4
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==3.4.7
cx-Freeze==6.5.3
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-allauth==0.44.0
django-appconf==1.0.4
django-autoslug==1.9.8
django-heroku==0.3.1
django-imagekit==4.0.2
django-registration==2.4.1
django-registration-bootstrap==0.2.0
django-rest-auth==0.9.5
django-social-share==2.2.1
django-stdimage==5.3.0
django-storages==1.11.1
django-taggit==1.4.0
django-utils-six==2.0
djangorestframework==3.12.4
Faker==8.1.0
future==0.18.2
httpie==2.4.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.7.3
incremental==21.3.0
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
mysql-connector-python==8.0.24
oauthlib==3.1.0
p5py==1.0.0
packaging==20.9
pefile==2019.4.18
pep517==0.10.0
pilkit==2.0
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.2
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.5
gunicorn==20.1.0
whitenoise==5.2.0
wrapt==1.12.1
